requirements.txt
numpy==1.21.6
Pillow==9.4.0
streamlit==1.28.0
matplotlib==3.7.3
reportlab==4.0.4
pypdf2==3.0.1